Also handle graveyard
