+++ src/main/java/com/minecolonies/coremod/colony/buildings/workerbuildings/BuildingGraveyard.java
@@ -3,7 +3,6 @@ package com.minecolonies.coremod.colony.buildings.workerbuildings;
 import com.ldtteam.blockout.views.Window;
 import com.minecolonies.api.blocks.AbstractBlockMinecoloniesNamedGrave;
 import com.minecolonies.api.blocks.ModBlocks;
-import com.minecolonies.api.colony.GraveData;
 import com.minecolonies.api.colony.ICitizenData;
 import com.minecolonies.api.colony.IColony;
 import com.minecolonies.api.colony.IColonyView;
@@ -13,13 +12,12 @@ import com.minecolonies.api.colony.buildings.workerbuildings.IBuildingPublicCraf
 import com.minecolonies.api.colony.jobs.IJob;
 import com.minecolonies.api.entity.citizen.Skill;
 import com.minecolonies.api.tileentities.TileEntityGrave;
-import com.minecolonies.api.tileentities.TileEntityNamedGrave;
 import com.minecolonies.api.util.BlockPosUtil;
 import com.minecolonies.api.util.ItemStackUtils;
 import com.minecolonies.api.util.Tuple;
 import com.minecolonies.api.util.WorldUtil;
 import com.minecolonies.api.util.constant.ToolType;
-import com.minecolonies.coremod.client.gui.huts.WindowHutGraveyardModule;
+import com.minecolonies.coremod.client.gui.huts.WindowHutWorkerModulePlaceholder;
 import com.minecolonies.coremod.colony.buildings.AbstractBuildingWorker;
 import com.minecolonies.coremod.colony.buildings.views.AbstractBuildingWorkerView;
 import com.minecolonies.coremod.colony.jobs.JobUndertaker;
@@ -27,20 +25,16 @@ import net.minecraft.block.BlockState;
 import net.minecraft.item.Items;
 import net.minecraft.nbt.CompoundNBT;
 import net.minecraft.nbt.ListNBT;
-import net.minecraft.nbt.StringNBT;
-import net.minecraft.network.PacketBuffer;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.Direction;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 import net.minecraftforge.common.util.Constants;
-import org.apache.commons.lang3.StringUtils;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.*;
 
-import static com.minecolonies.api.util.constant.Constants.TAG_STRING;
 import static com.minecolonies.api.util.constant.ToolLevelConstants.TOOL_LEVEL_WOOD_OR_GOLD;
 
 /**
@@ -64,11 +58,6 @@ public class BuildingGraveyard extends AbstractBuildingWorker implements IBuildi
     private static final int MAX_BUILDING_LEVEL = 5;
 
     /**
-     * NBTTag to store grave data.
-     */
-    private static final String TAG_GRAVE_DATA = "gravedata";
-
-    /**
      * NBTTag to store the visual grave positions.
      */
     private static final String TAG_VISUAL_GRAVES = "visualgraves";
@@ -89,26 +78,10 @@ public class BuildingGraveyard extends AbstractBuildingWorker implements IBuildi
     private static final String TAG_CURRENT_GRAVE = "currentGRAVE";
 
     /**
-     * The tag to store the list of resting citizen in this graveyard
-     */
-    private static final String TAG_RIP_CITIZEN_LIST = "ripCitizenList";
-
-    /**
-     * The list of resting citizen in this graveyard.
-     */
-    private final List<String> restingCitizen = new ArrayList<>();
-
-    /**
      * The grave the undertaker is currently collecting.
      */
     @Nullable
     private BlockPos currentGrave;
-
-    /**
-     * The data of the last grave dug by the undertaker.
-     */
-    @Nullable
-    private GraveData lastGraveData;
 
     /**
      * Grave positions
@@ -126,7 +99,6 @@ public class BuildingGraveyard extends AbstractBuildingWorker implements IBuildi
         super(c, l);
         keepX.put(itemStack -> ItemStackUtils.hasToolLevel(itemStack, ToolType.SHOVEL, TOOL_LEVEL_WOOD_OR_GOLD, getMaxToolLevel()), new net.minecraft.util.Tuple<>(1, true));
         keepX.put(itemStack -> itemStack.getItem() == Items.TOTEM_OF_UNDYING, new net.minecraft.util.Tuple<>(2, true));
-
     }
 
     /**
@@ -181,24 +153,6 @@ public class BuildingGraveyard extends AbstractBuildingWorker implements IBuildi
             currentGrave = BlockPosUtil.read(compound, TAG_CURRENT_GRAVE);
         }
 
-        restingCitizen.clear();
-        if (compound.getAllKeys().contains(TAG_RIP_CITIZEN_LIST))
-        {
-            final ListNBT ripCitizen = compound.getList(TAG_RIP_CITIZEN_LIST, TAG_STRING);
-            for (int i = 0; i < ripCitizen.size(); i++)
-            {
-                final String citizenName = ripCitizen.getString(i);
-                restingCitizen.add(citizenName);
-            }
-        }
-
-        if (compound.getAllKeys().contains(TAG_GRAVE_DATA))
-        {
-            lastGraveData = new GraveData();
-            lastGraveData.read(compound.getCompound(TAG_GRAVE_DATA));
-        }
-        else lastGraveData = null;
-
         visualGravePositions.clear();
         final ListNBT visualGraveTagList = compound.getList(TAG_VISUAL_GRAVES, Constants.NBT.TAG_COMPOUND);
         for (int i = 0; i < visualGraveTagList.size(); ++i)
@@ -218,18 +172,6 @@ public class BuildingGraveyard extends AbstractBuildingWorker implements IBuildi
         if (currentGrave != null)
         {
             BlockPosUtil.write(compound, TAG_CURRENT_GRAVE, currentGrave);
-        }
-
-        @NotNull final ListNBT ripCitizen = new ListNBT();
-        for (@NotNull final String citizenName : restingCitizen)
-        {
-            ripCitizen.add(StringNBT.valueOf(citizenName));
-        }
-        compound.put(TAG_RIP_CITIZEN_LIST, ripCitizen);
-
-        if(lastGraveData != null)
-        {
-            compound.put(TAG_GRAVE_DATA, lastGraveData.write());
         }
 
         @NotNull final ListNBT visualGraveTagList = new ListNBT();
@@ -281,90 +223,6 @@ public class BuildingGraveyard extends AbstractBuildingWorker implements IBuildi
         return MAX_BUILDING_LEVEL;
     }
 
-    /**
-     * Method to serialize data to send it to the view.
-     *
-     * @param buf the used ByteBuffer.
-     */
-    @Override
-    public void serializeToView(@NotNull final PacketBuffer buf)
-    {
-        super.serializeToView(buf);
-
-        final List<BlockPos> graves = new ArrayList<>(colony.getGraveManager().getGraves().keySet());
-        final List<BlockPos> cleanList = new ArrayList<>();
-
-        for (@NotNull final BlockPos grave : graves)
-        {
-            if (WorldUtil.isBlockLoaded(colony.getWorld(), grave))
-            {
-                final TileEntity tileEntity = getColony().getWorld().getBlockEntity(grave);
-                if (tileEntity instanceof TileEntityGrave)
-                {
-                    cleanList.add(grave);
-                }
-            }
-        }
-
-        // grave list
-        buf.writeInt(cleanList.size());
-        for (@NotNull final BlockPos grave : cleanList)
-        {
-            buf.writeBlockPos(grave);
-        }
-
-        //resting citizen list
-        buf.writeInt(restingCitizen.size());
-        for (@NotNull final String citizenName : restingCitizen)
-        {
-            buf.writeUtf(citizenName);
-        }
-        buf.writeInt(restingCitizen.size());
-    }
-
-    /**
-     * Add a citizen to the list of resting citizen in this graveyard
-     */
-    public void buryCitizenHere(final Tuple<BlockPos, Direction> positionAndDirection)
-    {
-        if(lastGraveData != null && !restingCitizen.contains(lastGraveData.getCitizenName()))
-        {
-            Direction facing = positionAndDirection.getB();
-            if(facing == Direction.UP || facing == Direction.DOWN)
-            {
-                facing = Direction.NORTH; //prevent setting an invalid HorizontalDirection
-            }
-            getColony().getWorld().setBlockAndUpdate(positionAndDirection.getA(), ModBlocks.blockNamedGrave.defaultBlockState().setValue(AbstractBlockMinecoloniesNamedGrave.FACING, facing));
-
-            TileEntity tileEntity = getColony().getWorld().getBlockEntity(positionAndDirection.getA());
-            if (tileEntity instanceof TileEntityNamedGrave)
-            {
-                final String firstName = StringUtils.split(lastGraveData.getCitizenName())[0];
-                final String lastName = lastGraveData.getCitizenName().replaceFirst(firstName,"");
-
-                final ArrayList<String> lines = new ArrayList<>();
-                lines.add(firstName);
-                lines.add(lastName);
-                if (lastGraveData.getCitizenJobName() != null) { lines.add(lastGraveData.getCitizenJobName()); }
-                ((TileEntityNamedGrave) tileEntity).setTextLines(lines);
-            }
-
-            restingCitizen.add(lastGraveData.getCitizenName());
-            markDirty();
-        }
-    }
-
-    public void setLastGraveData(final GraveData graveData)
-    {
-        this.lastGraveData = graveData;
-        markDirty();
-    }
-
-    public GraveData getLastGraveData()
-    {
-        return this.lastGraveData;
-    }
-
     @Override
     public void registerBlockPosition(@NotNull final BlockState state, @NotNull final BlockPos pos, @NotNull final World world)
     {
@@ -409,18 +267,6 @@ public class BuildingGraveyard extends AbstractBuildingWorker implements IBuildi
     public static class View extends AbstractBuildingWorkerView
     {
         /**
-         * Contains a view object of all the graves in the colony.
-         */
-        @NotNull
-        private List<BlockPos> graves = new ArrayList<>();
-
-        /**
-         * Contains a view object of all the restingCitizen in the colony.
-         */
-        @NotNull
-        private List<String> restingCitizen = new ArrayList<>();
-
-        /**
          * Public constructor of the view, creates an instance of it.
          *
          * @param c the colony.
@@ -435,50 +281,7 @@ public class BuildingGraveyard extends AbstractBuildingWorker implements IBuildi
         @NotNull
         public Window getWindow()
         {
-            return new WindowHutGraveyardModule(this);
-        }
-
-        @Override
-        public void deserialize(@NotNull final PacketBuffer buf)
-        {
-            super.deserialize(buf);
-
-            graves = new ArrayList<>();
-            final int size = buf.readInt();
-            for (int i = 1; i <= size; i++)
-            {
-                @NotNull final BlockPos pos = buf.readBlockPos();
-                graves.add(pos);
-            }
-
-            restingCitizen = new ArrayList<>();
-            final int sizeRIP = buf.readInt();
-            for (int i = 1; i <= sizeRIP; i++)
-            {
-                restingCitizen.add(buf.readUtf());
-            }
-        }
-
-        /**
-         * Getter of the graves list.
-         *
-         * @return an unmodifiable List.
-         */
-        @NotNull
-        public List<BlockPos> getGraves()
-        {
-            return Collections.unmodifiableList(graves);
-        }
-
-        /**
-         * Getter of the restingCitizen list.
-         *
-         * @return an unmodifiable List.
-         */
-        @NotNull
-        public List<String> getRestingCitizen()
-        {
-            return Collections.unmodifiableList(restingCitizen);
+            return new WindowHutWorkerModulePlaceholder<>(this, GRAVEYARD);
         }
     }
 }
